Improve day 7 solution
Reduce code duplication and get rid of all the unwrap() calls in the
compare function. We can now also get back to the Ord implementation.
This is accomplished by

a) introducing the joker alongside the jack ('J') by patching the hands
   before running part 2. This removes a lot of code duplication as we
   can now use the same Ord implementation for both parts.
b) introducing the Card enum. This allows us to get rid of error
   handling in the kind() function because we no longer need to account
   for arbitrary characters (i.e. cards) in our hands since we already
   did in the parsing step.